fix authentication issues by update firabse authentication and remove password field from real time
