+++ app/src/main/java/com/example/project_yougo/feed/AddCommentFragment.java
@@ -14,7 +14,7 @@ import android.widget.EditText;
 import android.widget.Toast;
 
 import com.example.project_yougo.R;
-import com.example.project_yougo.model.UserModelFirebase;
+import com.example.project_yougo.model.user.UserModelFirebase;
 import com.example.project_yougo.model.comment.CommentModel;
 
 /**
