+++ app/src/main/java/com/example/project_yougo/feed/CommentListFragment.java
@@ -16,8 +16,8 @@ import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.example.project_yougo.R;
-import com.example.project_yougo.model.User;
-import com.example.project_yougo.model.UserModelFirebase;
+import com.example.project_yougo.model.user.User;
+import com.example.project_yougo.model.user.UserModelFirebase;
 import com.example.project_yougo.model.comment.CommentModel;
 import com.example.project_yougo.model.post.Comment;
